feat: add negative_transfer vulnerable/secure contract with tests
- Vulnerable TokenContract: transfer() accepts negative amounts, reversing
  the transfer direction (sender gains tokens instead of sending them)
- Secure mirror: SecureTokenContract asserts amount > 0 before mutating balances
- Tests: positive transfer, negative amount exploit, secure rejects negative/zero
